+++ app/Modules/Setup/Views/setup/xferaccount.blade.php
@@ -23,7 +23,7 @@
                         <br><br><br>
                         {!! Form::text('olddbname') !!}
                         <br><br><br>
-                        {!! Form::submit(trans('fi.i_accept'), ['class' => 'btn btn-primary']) !!}
+                        {!! Form::submit(trans('fi.continue'), ['class' => 'btn btn-primary']) !!}
                     </div>
 
                 </div>